Update #1: Change the return value for the "HourlyForecastContainer" component so that it returns a "section" element with an id of "hourly-forecast-container" and no content.
Update #2: Define a type aliases for the data being
displayed in the "HourlyForecastContainer" component.

Update #3: Import "useState" and use it to define an
array of HourForecasts as part of the component's local
state and give it a temporary value.

Update #4: Without actually defining the logic of the
function, define a function for refreshing hour
forecasts and give it a single paramater which is a
callback function for refreshing the info of a single
hour forecast.

Update #5: Define a temporary forecast refresh method
to use when the "ForcastInfoType" type alias is set to
"number".