Update #1: Update "_hourlyForecast.scss" so that the "#hourly-forecast-container" element fits most of the "#weather-hourly-forecast" element.
Update #2: Update "_hourlyForecast.scss" so that
the "#hourly-forecast-container" element contains
5 numbered columns that fit most of the element and
are evenly spaced within the element.

Update #3: Update "_hourlyForecast.scss" so that,
on viewports < 768 pixels, the width for the
columns of the "#hourly-forecast-container" element
cause the element to have overflow.

Update #4: Update "_hourlyForecast.scss" so that,
on viewports < 768 pixels, the
"#hourly-forecast-container" element has horizontal
scrolling, a scroll behavior of "smooth", and it's
"scroll-snap-align" property is set to "center".