Update #1: Add 3 sections and a footer to the columns of the "HourlyForecastContainer" component.
Update #2: Add an "h4" element and a "p" element
to the new sections added.

Update #3: Add an "h4" element to the footer added.

Update #4: In the "houly-info" sections define a
label and value for hourly temperature, humidity,
and wind.

Update #5: Define a placeholder time (using only
the hour and AM/PM) for the "h4" element of the
footer added.