Update #1: Update the "handleClick" function so that it accepts a 2nd parameter when called (the index of the list item being clicked).
Update #2: Update the "handleClick" function so
that the only thing it does is update the value
of "currentSelection" in the redux store to the
index of the list item being clicked on, if the
list item being clicked on isn't already
selected.

Update #3: Give the list items of the
"WeatherRequestHistory" component a conditional
value for it's "className" attribute where the
value is "selected" when the index of the list
item matches the value of "currentSelection" in
the redux store and, for all other situations,
the value of the "className" attribute is "".