Update #1: Turn the left column of the "#main-info" element into a flexbox and give it a width of 50%.
Update #2: Set the "justify-content"
property of the "#temp-info" element to
"space-between".

Update #3: Give the "#current-temp-label"
element a "grey" background and set the
"padding" property to ".5rem".

Update #4: Make it so that the "#temp-info"
and "#condition-info" elements take up 50%
of the width inside the left column of the
"main-info" element.

Update #5: Style the "#condition-info"
element similarly to the "#temp-info"
element.