Update #1: Update the style for the "#location-val" element so that:
1. The font-size is "1.2rem".
2. The element has a padding of
".25rem".
3. the background of the element is
"red".
4. The text sits at the center of the
element.

Update #2: Simplify the styles for the columns
of the "#main-info" element.

Update #3: Remove the style for the "#wind-info"
element.

Update #4: Update the style for the columns
of the "#main-info" element so that the
columns have a width of 100% on mobile
devices and a width of 50% on all other devices.

Update #5: Update the style for the "#time-bar"
element so that the font-size is "1.1rem" on
mobile devices and "1.4rem on all other devices".

Update #6: Clean up the code in
"_weatherHeader.scss".