Update #1: Define a new function "getCurrentHour" so that it has 2 parameters, "selectedInfo" of type any and "currentDate" of type date, and returns the hour of the day for the date given by "currentDate".
Update #2: Add another parameter to the
"getHourlyForecasts" function for "selectedInfo".

Update #3: Add a new "hourly-info" section to each
hour forecast for weather condition.

Update #4: Add code that will be used to access
the hourly forecasts for a selected city based on
the current time for the selected city and the
current day of month.