Update #1: Rename the "InterfaceMessages" component to "MessageOverlay".
Update #2: Replace the message windows
with a single message window that can
be styled by swapping out different
class names.

Update #3: Create a function called
"toggleMessageOverlay" that can be used
to show/hide the "MessageOverlay"
component.

Update #4: Create a function called
"toggleMessageWindow" that can be used
to change whether or not the message
window is active and being displayed on
the "MessageOverlay" component.

Update #5: Create a function called
"setWindowType" that can be used to swap
the appearance of the message window using
any of the defined classes.

Update #6: Create a demonstration using
the newly created functions that shows how
the "MessageOverlay" component might work.